AlgorithmDataStructure/OnlineJudge/f548a99304e14891ba985d653db34eef/178-datastructure_chapter4_1/src/test/java/BinTreeTest.java
import org.junit.Test;
import static org.junit.Assert.*;

public class BinTreeTest {
    @Test
    public void test1(){

        BinTree root = BinTree.CreateBT("ABDCEGFHI".toCharArray(), "DBAEGCHFI".toCharArray());
        assertTrue("ABDCEGFHI".equals(BinTree.PreOrderTraversal(root).toString()));
        assertTrue("DBAEGCHFI".equals(BinTree.InOrderTraversal(root).toString()));
        assertTrue("DBGEHIFCA".equals(BinTree.PostOrderTraversal(root).toString()));
    }

    @Test
    public void test2(){
        BinTree root = BinTree.CreateBT("ABEFCGDKIHJ".toCharArray(), "EFBGCDAIJHK".toCharArray());
        assertTrue("ABEFCGDKIHJ".equals(BinTree.PreOrderTraversal(root).toString()));
        assertTrue("EFBGCDAIJHK".equals(BinTree.InOrderTraversal(root).toString()));
        assertTrue("FEGDCBJHIKA".equals(BinTree.PostOrderTraversal(root).toString()));
    }

}
